ci: Fix CI issues with different combinations of exprs
