+++ datafusion/core/src/datasource/listing/helpers.rs
@@ -23,7 +23,6 @@ use std::sync::Arc;
 
 use super::ListingTableUrl;
 use super::PartitionedFile;
-use crate::datasource::physical_plan::parquet::would_column_prevent_pushdown;
 use crate::execution::context::SessionState;
 use datafusion_common::{Result, ScalarValue};
 use datafusion_expr::{BinaryExpr, Operator};
@@ -116,31 +115,6 @@ pub fn expr_applicable_for_cols(col_names: &[&str], expr: &Expr) -> bool {
     })
     .unwrap();
     is_applicable
-}
-
-/// recurses through expr as a trea, finds all `column`s, and checks if any of them would prevent
-/// this expression from being predicate pushed down. If any of them would, this returns false.
-/// Otherwise, true.
-pub fn can_expr_be_pushed_down_with_schemas(
-    expr: &Expr,
-    file_schema: &Schema,
-    table_schema: &Schema,
-) -> bool {
-    let mut can_be_pushed = true;
-    expr.apply(|expr| match expr {
-        Expr::Column(column) => {
-            can_be_pushed &=
-                !would_column_prevent_pushdown(column.name(), file_schema, table_schema);
-            Ok(if can_be_pushed {
-                TreeNodeRecursion::Jump
-            } else {
-                TreeNodeRecursion::Stop
-            })
-        }
-        _ => Ok(TreeNodeRecursion::Continue), // we never return an Err, so we can safely unwrap this
-    })
-    .unwrap();
-    can_be_pushed
 }
 
 /// The maximum number of concurrent listing requests
