+++ datafusion/core/src/datasource/listing/mod.rs
@@ -31,7 +31,6 @@ use std::pin::Pin;
 use std::sync::Arc;
 
 pub use self::url::ListingTableUrl;
-pub use helpers::can_expr_be_pushed_down_with_schemas;
 pub use table::{ListingOptions, ListingTable, ListingTableConfig};
 
 /// Stream of files get listed from object store
